MIM-11 修复平台预览 CI 门禁
