+++ ios/MimiRemote/Sources/State/AppStore.swift
@@ -376,7 +376,10 @@ final class AppStore: ObservableObject {
         if let previewPlatform = debugLaunchConfiguration.hostPlatformPreview {
             let previewProfile = ConnectionProfile(
                 id: "debug-platform-preview",
-                displayName: "\(previewPlatform.displayName ?? "Unknown") 主机",
+                displayName: L10n.format(
+                    "ui.host_platform_preview_name",
+                    previewPlatform.displayName ?? L10n.text("ui.unknown")
+                ),
                 endpoint: "http://127.0.0.1:8787",
                 lastSuccessfulAt: Date(),
                 installationID: "debug-platform-preview",
@@ -386,7 +389,7 @@ final class AppStore: ObservableObject {
                 previewProfile,
                 ConnectionProfile(
                     id: "debug-platform-companion",
-                    displayName: "另一台主机",
+                    displayName: L10n.text("ui.another_host"),
                     endpoint: "http://127.0.0.1:8788",
                     lastSuccessfulAt: nil,
                     installationID: "debug-platform-companion",
@@ -1981,70 +1984,3 @@ final class AppStore: ObservableObject {
         }
     }
 }
-
-#if DEBUG
-private struct DebugLaunchConfiguration {
-    let opensWorkbenchWithoutPairing: Bool
-    let seedsWorkbenchUI: Bool
-    let seedsQueuedTurnsUI: Bool
-    let seedsMCPApprovalUI: Bool
-    let hostPlatformPreview: HostPlatform?
-    let endpoint: String?
-    let token: String?
-
-    static func current(processInfo: ProcessInfo = .processInfo) -> DebugLaunchConfiguration {
-        let arguments = processInfo.arguments
-        let environment = processInfo.environment
-        let seedsQueuedTurnsUI = arguments.contains("--debug-seed-queue-ui")
-            || boolValue(environment["MIMI_DEBUG_SEED_QUEUE_UI"])
-        let seedsMCPApprovalUI = arguments.contains("--debug-seed-mcp-approval-ui")
-            || boolValue(environment["MIMI_DEBUG_SEED_MCP_APPROVAL_UI"])
-        let hostPlatformPreview = argumentValue(named: "--debug-host-platform", in: arguments)
-            .map { HostPlatform(serverValue: $0) }
-        return DebugLaunchConfiguration(
-            opensWorkbenchWithoutPairing: arguments.contains("--debug-skip-pairing")
-                || boolValue(environment["MIMI_DEBUG_SKIP_PAIRING"])
-                || hostPlatformPreview != nil,
-            seedsWorkbenchUI: arguments.contains("--debug-seed-ui")
-                || boolValue(environment["MIMI_DEBUG_SEED_UI"])
-                || seedsQueuedTurnsUI
-                || seedsMCPApprovalUI,
-            seedsQueuedTurnsUI: seedsQueuedTurnsUI,
-            seedsMCPApprovalUI: seedsMCPApprovalUI,
-            hostPlatformPreview: hostPlatformPreview,
-            endpoint: argumentValue(named: "--debug-endpoint", in: arguments)
-                ?? environment["MIMI_DEBUG_ENDPOINT"],
-            token: argumentValue(named: "--debug-token", in: arguments)
-                ?? environment["MIMI_DEBUG_TOKEN"]
-        )
-    }
-
-    private static func argumentValue(named name: String, in arguments: [String]) -> String? {
-        let inlinePrefix = "\(name)="
-        if let inlineValue = arguments.first(where: { $0.hasPrefix(inlinePrefix) }) {
-            let value = String(inlineValue.dropFirst(inlinePrefix.count))
-                .trimmingCharacters(in: .whitespacesAndNewlines)
-            return value.isEmpty ? nil : value
-        }
-        guard let index = arguments.firstIndex(of: name) else {
-            return nil
-        }
-        let valueIndex = arguments.index(after: index)
-        guard arguments.indices.contains(valueIndex),
-              !arguments[valueIndex].hasPrefix("--") else {
-            return nil
-        }
-        let value = arguments[valueIndex].trimmingCharacters(in: .whitespacesAndNewlines)
-        return value.isEmpty ? nil : value
-    }
-
-    private static func boolValue(_ rawValue: String?) -> Bool {
-        switch rawValue?.trimmingCharacters(in: .whitespacesAndNewlines).lowercased() {
-        case "1", "true", "yes", "y", "on":
-            return true
-        default:
-            return false
-        }
-    }
-}
-#endif
